+++ services/server/src/server/server.ts
@@ -22,7 +22,7 @@ import logger from "../common/logger";
 import routes from "./routes";
 import genericErrorHandler from "../common/errors/GenericErrorHandler";
 import { validateAddresses, validateSingleAddress } from "./common";
-import { initDeprecatedRoutes } from "./deprecated.routes";
+import { initDeprecatedRoutes } from "./apiv1/deprecated.routes";
 import getSessionMiddleware from "./session";
 import { Services } from "./services/services";
 import { StorageServiceOptions } from "./services/StorageService";
